Cleanup iOS proxy configuration
